move legacy code to legacy package
