+++ common/geo.go
@@ -1,7 +1,21 @@
 package common
 
 import (
+	"errors"
+	"os"
+	"regexp"
 	"time"
+)
+
+var (
+	// This is the bucket containing maxmind files.
+	MaxmindBucketName = "downloader-" + os.Getenv("GCLOUD_PROJECT")
+	// This is the regex used to filter for which files we want to consider acceptable for using with Geolite2
+	GeoLite2Regex = regexp.MustCompile(`Maxmind/\d{4}/\d{2}/\d{2}/\d{8}T\d{6}Z-GeoLite2-City-CSV\.zip`)
+)
+
+const (
+	MaxmindPrefix = "Maxmind/" // Folder containing the maxmind files
 )
 
 // The GeolocationIP struct contains all the information needed for the
@@ -40,4 +54,16 @@ type RequestData struct {
 	IP        string    // Holds the IP from an incoming request
 	IPFormat  int       // Holds the ip format, 4 or 6
 	Timestamp time.Time // Holds the timestamp from an incoming request
+}
+
+// ExtractDateFromFilename return the date for a filename like
+// gs://downloader-mlab-oti/Maxmind/2017/05/08/20170508T080000Z-GeoLiteCity.dat.gz
+// TODO move this to maxmind package
+func ExtractDateFromFilename(filename string) (time.Time, error) {
+	re := regexp.MustCompile(`[0-9]{8}T`)
+	filedate := re.FindAllString(filename, -1)
+	if len(filedate) != 1 {
+		return time.Time{}, errors.New("cannot extract date from input filename")
+	}
+	return time.Parse(time.RFC3339, filedate[0][0:4]+"-"+filedate[0][4:6]+"-"+filedate[0][6:8]+"T00:00:00Z")
 }
