+++ main.go
@@ -7,6 +7,7 @@ import (
 	"os"
 
 	"github.com/m-lab/annotation-service/handler"
+	"github.com/m-lab/annotation-service/legacy"
 	"github.com/m-lab/annotation-service/metrics"
 )
 
@@ -14,7 +15,7 @@ import (
 func updateMaxmindDatasets(w http.ResponseWriter, r *http.Request) {
 	log.Printf("Update the list of maxmind datasets.\n")
 
-	err := handler.UpdateFilenamelist("downloader-" + os.Getenv("GCLOUD_PROJECT"))
+	err := legacy.UpdateFilenamelist("downloader-" + os.Getenv("GCLOUD_PROJECT"))
 	if err != nil {
 		log.Print(err.Error())
 		http.Error(w, err.Error(), http.StatusInternalServerError)
